// frontend/src/api/types/CreateTournament.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TournamentState } from "./TournamentState";

export type CreateTournament = { name: string, start_date: string, end_date: string | null, state: TournamentState, win_team_id: string | null, };